Fix some errors #1

+++ src/rollercoastertycoon/accident/Accident.java
@@ -3,7 +3,7 @@ package rollercoastertycoon.accident;
 public interface Accident {
     void happen();
     String getMessage();
-    int getCost();
+    long getCost();
     int getVisitorDecreasePercentage();
     float getProbability();
     boolean isAbleToHapen();
